re-structured the main.py added support for solution explanation
