+++ ta-scheduler-timefold/pyproject.toml
@@ -8,7 +8,7 @@ name = "hello_world"
 version = "1.0.0"
 requires-python = ">=3.10"
 dependencies = [
-    'timefold == 1.17.0b0',
+    'timefold >= 1.17.0b0',
     'pytest == 8.2.2'
 ]
