+++ ta-scheduler-timefold/src/hello_world/utils.py
@@ -298,7 +298,6 @@ class DataConstructor:
 #         print(f"Report has been saved to {filename}")
 
 
-"""Provides printing utilities for the assignment matrix."""
 def id_generator():
     """Generates unique IDs for the shift assignments."""
     current = 0
